test: split test_app file
