+++ tests/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+from flaskr.app import app
+
+
+@pytest.fixture
+def client():
+    return app.test_client()
